+++ src/index.js
@@ -1,5 +1,5 @@
-import UserInputModel from "./input-checker/model.js";
-import UserInputView from "./input-checker/view.js";
+import UserInputModel from "./user-input/model.js";
+import UserInputView from "./user-input/view.js";
 
 import GameRuleModel from "./game-manager/model.js";
 import GameRuleView from "./game-manager/view.js";
