+++ src/user-input/model.js
