+++ src/user-input/view.js
